% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_age_gender_bar.R
\name{AER_age_gender_bar}
\alias{AER_age_gender_bar}
\title{AER age-gender bar}
\usage{
AER_age_gender_bar(data, xvar, yvar, group, fill_color1 = "#65B32E",
  fill_color2 = "#7CBDC4", ytitle = "Rate")
}
\arguments{
\item{data}{Your data frame}

\item{xvar}{Variable on the x-axis in quotes; e.g. age group}

\item{yvar}{Variable on the y-axis in quotes with the rate or number of cases}

\item{group}{A grouping variable in quotes, e.g. gender as in the AER.}

\item{fill_color1}{Bar 1 colour; defaults to ECDC green "#65B32E"}

\item{fill_color2}{Bar 2 colour; defaults to light blue "#7CBDC1"}

\item{ytitle}{y-axis title; defaults to "Rate".}
}
\description{
This function draws a barchart by age group and gender(or possibly other grouping), AER style. 
Expects aggregated data.
}
\examples{
# Create dummy data
mydat <- data.frame(Gender=c("F", "F", "M", "M"), 
AgeGroup = c("0-65", "65+", "0-65", "65+"), 
NumberOfCases = c(54,43,32,41))

# Plot the dummy data
AER_age_gender_bar(mydat, xvar = AgeGroup, 
yvar = NumberOfCases, 
group = Gender, 
ytitle = "Number of cases")
}
\keyword{AER_age_gender_bar}
