% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_seasonality_graph.R
\name{AER_seasonality_graph}
\alias{AER_seasonality_graph}
\title{AER}
\usage{
AER_seasonality_graph(data, xvar = "timecode", yvar = "yvalue",
  yearvar = "timeyear", monthvar = "timemonth", year = 2015)
}
\arguments{
\item{data}{Your data frame}

\item{xvar}{Variable on the x-axis in quotes; e.g. age group}

\item{yvar}{Variable on the y-axis in quotes with the rate or number of cases}

\item{yearvar}{Var...}

\item{monthvar}{Var...}

\item{year}{Var...}
}
\description{
This function ...
}
\keyword{AER_seasonality_graph}
