% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AERdata}
\alias{AERdata}
\title{Dataset for AER production}
\format{A data frame with 90,649 rows and 9 variables:
\describe{
  \item{HealthTopic}{Disease name e.g. Botulism, Cholera, etc}
  \item{Population}{Population characteristics e.g. All cases, Confirmed cases, etc}
  \item{Indicator}{Indicator e.g. Age-standardised rate, Hospitalised cases, Reported cases, Number of deaths, etc.}
  \item{Unit}{Unit of the indicator e.g. N \% or 1/100000}
  \item{Time}{Time variable including both yearly data from 1995 to 2016, and monthly data from 1998-01 to 2016-12}
  \item{RegionCode}{Geographical level including country codes e.g. AT, BE, BG, etc.}
  \item{RegionName}{Geographical level including country names e.g. Austria, Belgium, Bulgaria, etc.}
  \item{NumValue}{Value of the indicator}
  \item{TxtValue}{...}
}}
\source{
\url{http://atlas.ecdc.europa.eu/public/index.aspx}
}
\usage{
data(AERdata)
}
\description{
A dummy dataset (ECDC Atlas export) containing the information required to build the AER
for a set of 6 diseases: Botulism, Cholera, Crimean-Congo haemorrhagic fever,
Pertussis, Salmonellosis, Tuberculosis
}
\keyword{datasets}
