% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AERparams}
\alias{AERparams}
\title{Dataset including parameters for AER production}
\format{A data frame with 53 rows (corresponding to the 53 diseases) and 28 variables:
\describe{
  \item{HealthTopic}{Disease code e.g. ANTH, SALM, etc.}
  \item{DG}{(optional) Disease group e.g. FWD}
  \item{DP}{(optional) Disease programme e.g. FWD}
  \item{Label}{Disease label e.g. salmonellosis, anthrax}
  \item{FrequencyCategory}{(optional) Frequency of the disease e.g. VERY RARE, NON-RARE, etc.}
  \item{MeasurePopulation}{Type of population presented for this disease e.g. all or confirmed cases}
  \item{DatePublicAtlas}{Date of latest availability in the public access of the Atlas}
  \item{TableUse}{Type of table to present in the report e.g. NO table, ASR, RATE table, etc.}
  \item{TableRatesLabel}{Label to use in the table for presenting rates}
  \item{TableRatesNoDecimals}{Number of decimals to use for presenting rates}
  \item{TableASRNoDecimals}{Number of descimals to use for presenting ASR}
  \item{AgeGenderUse}{Type of age and gender bar graph to present e.g. NO graph, AG-RATE, AG-COUNT, etc.}
  \item{AgeGenderBarGraphLabel}{Label to use in the age and gender bar graph}
  \item{AgeGenderGraphNoDecimals}{Number of decimals to use for presenting rates in the age an dgender bar graph}
  \item{TSTrendGraphUse}{Logical Y/N for trend line graph inclusion}
  \item{TSSeasonalityGraphUse}{Logical Y/N for seasonal line graph inclusion}
  \item{TSSpecific}{Logical Y/N for specific line graph inclusion}
  \item{MapNumbersUse}{Logical Y/N for map inclusion}
  \item{MapRatesUse}{Logical Y/N for map inclusion}
  \item{MapRatesNoDecimals}{(optional) Number of decimals to use for presenting maps}
  \item{MapASRUse}{Logical Y/N for map inclusion}
  \item{MapASRNoDecimals}{(optional) Number of decimals to use for presenting maps}
  \item{Transmission}{Not implemented yet}
  \item{TransmissionNoDecimals}{Not implemented yet}
}}
\usage{
AERparams
}
\description{
A dataset containing the parameters required to build the AER
for all 53 diseases included in TESSy
}
\keyword{datasets}
