% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{AgeGenderBarGph}
\alias{AgeGenderBarGph}
\title{AgeGender}
\usage{
AgeGenderBarGph(disease, year, x)
}
\arguments{
\item{disease}{character string for the health topic}

\item{year}{numeric}

\item{x}{dataset including required data for AE}
}
\value{
A table
}
\description{
Function drawing ....
}
