% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_table.R
\name{cleanECDCTable}
\alias{cleanECDCTable}
\title{Cleaning the final table}
\usage{
cleanECDCTable(x, Country, GeoCode)
}
\arguments{
\item{x}{dataset to clean}

\item{Country}{character vector, geo code of each Member State
(default EpiReport::MSCode$Country)}

\item{GeoCode}{character vector, corresponding geo labels of each Member State
(default EpiReport::MSCode$GeoCode)}
}
\value{
cleaned ECDC table
}
\description{
Cleaning the final table: identifying missing reports with '-',
replacing the Member State codes with Member State names,
identifying not reporting Member States with '.'
}
