% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_tables.R
\name{cleaningECDCTable}
\alias{cleaningECDCTable}
\title{Cleaning the final table}
\usage{
cleaningECDCTable(x, MSCode)
}
\arguments{
\item{x}{dataset including required data for AER}

\item{MSCode}{dataset containing the correspondence table of geo code and labels}
}
\value{
cleaned ECDC table
}
\description{
Cleaning the final table: identifying missing reports with '-',
replacing the Member State codes with Member State names,
identifying not reporting Member States with '.'
}
