% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{cleanMeasureCode}
\alias{cleanMeasureCode}
\title{Clean the MeasureCode variable}
\usage{
cleanMeasureCode(var)
}
\arguments{
\item{var}{vector variable, variable to clean}
}
\value{
cleaned vector variable
}
\description{
Clean the MeasureCode variable and replace the specific codes with the generic ones
(e.g. ACCUTE.AGE_GENDER.RATE will be replaced by CONFIRMED.AGE_GENDER.RATE)
}
\examples{
x <- EpiReport::SALM2016
x$MeasureCode <- cleanMeasureCode(x$MeasureCode)
}
\seealso{
\code{\link{SALM2016}}
}
