% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{filterDisease}
\alias{filterDisease}
\title{Filter disease parameters}
\usage{
filterDisease(dis, reportParameters)
}
\arguments{
\item{dis}{character string containing disease code}

\item{reportParameters}{dataset of parameters for the report}
}
\value{
row corresponding to the parameters of the selected disease
}
\description{
Filter disease parameters
}
