% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_report.R
\name{getAER}
\alias{getAER}
\title{Produce the AER report}
\usage{
getAER(template, outputPath, x, disease, year, reportParameters, pathPNG)
}
\arguments{
\item{template}{The word document in which to add the table.
Default value is the empty template included in the package getTemplate().}

\item{outputPath}{The path for the word output.
Default value is the current working directory getwd().}

\item{x}{dataset including required data for AER}

\item{disease}{character string, disease name}

\item{year}{numeric, year to produce the report for}

\item{reportParameters}{dataset of parameters for the report}

\item{pathPNG}{character string, path to the folder containing the maps in PNG}
}
\value{
A word document
}
\description{
Function to generate the AER Word report
}
