% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_report.R
\name{getTemplate}
\alias{getTemplate}
\title{Get AER Word template}
\usage{
getTemplate(output_path)
}
\arguments{
\item{output_path}{character string, the full path where to create the Word output.
Defaut location will be the current working directory (default \code{getwd()})}
}
\value{
A word document
}
\description{
Function to export the generic Microsoft Word template for the ECDC
annual epidemiological report (AER) \cr
(see reports already available on the ECDC dedicated web page
\url{https://ecdc.europa.eu/en/annual-epidemiological-reports})
}
\examples{
# getTemplate(output_path = "C:/R/AER")

}
