% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_trend_graph.R
\name{getTrend}
\alias{getTrend}
\title{Get the trend of reported/confirmed cases graph}
\usage{
getTrend(x, disease = "SALM", year = 2016, reportParameters, MSCode,
  index = 1, doc)
}
\arguments{
\item{x}{dataset}

\item{disease}{character string, disease name}

\item{year}{numeric, year to produce the report for}

\item{reportParameters}{dataset of parameters for the report}

\item{MSCode}{dataset containing the correspondence table of geo code and labels}

\item{index}{figure number}

\item{doc}{Word document}
}
\value{
Word doc a ggplot2 preview
}
\description{
Function returning the plot
}
