% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_maps.R
\name{includemap}
\alias{includemap}
\title{Function including the PNG map}
\usage{
includemap(disease, year, reportParameters, index, pathPNG, doc, pop,
  namePNGsuffix, unit, mapBookmark, captionBookmark)
}
\arguments{
\item{disease}{character string, disease name}

\item{year}{numeric, year to produce the report for}

\item{reportParameters}{dataset of parameters for the report}

\item{index}{figure number}

\item{pathPNG}{character string, path to the folder containing the maps in PNG}

\item{doc}{Word document}

\item{pop}{Label of the type of population used in the caption}

\item{namePNGsuffix}{Suffix of the PNG file name of the map}

\item{unit}{Label of the unit used in the caption}

\item{mapBookmark}{Bookmark for the map in the Word document}

\item{captionBookmark}{Bookmark for the caption in the Word document}
}
\value{
Word doc a preview
}
\description{
Function including the PNG map
}
