% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSCode}
\alias{MSCode}
\title{Dataset correspondence table for country names and code}
\format{A data frame with 32 rows and 2 variables:
\describe{
  \item{Country}{Country name}
  \item{GeoCode}{Country code}
}}
\usage{
data(MSCode)
}
\description{
Dataset correspondence table for country names and code
}
\keyword{datasets}
