% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_quasinum.R
\name{order_quasinum}
\alias{order_quasinum}
\title{A function to order 'quasinumerical' (i.e. categorical with values such as "15-30" or "<18") integer vectors into
increasing order. Currently handles away the following non-numerical characters "-", ">", "<", ">=", "<=", "+".}
\usage{
order_quasinum(x)
}
\arguments{
\item{x}{A character vector with 'quasinumerical' values}
}
\description{
A function to order 'quasinumerical' (i.e. categorical with values such as "15-30" or "<18") integer vectors into
increasing order. Currently handles away the following non-numerical characters "-", ">", "<", ">=", "<=", "+".
}
\examples{
age1 <- c("<1", "1-15", "16-24", ">65", "25-65")
age2 <- c("0-4", "5-10", ">65", "25-64", "11-25")
age3 <- c("5-10", ">65", "25-64", "11-25", "<=4")
age4 <- c(">=65", "<18", "18-64")
age5 <- c("5-10", "+65", "25-64", "11-25", "0-4")

age1
order_quasinum(age1)
age2
order_quasinum(age2)
age3
order_quasinum(age3)
age4
order_quasinum(age4)

age5
order_quasinum(age5)
}
\author{
Tommi Karki
}
\keyword{order}
