% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_age_gender_bar.R
\name{plotAge}
\alias{plotAge}
\title{AER age bar graph}
\usage{
plotAge(data, xvar = "XLabel", yvar = "YValue",
  fill_color1 = "#65B32E", ytitle = "Rate")
}
\arguments{
\item{data}{Your data frame}

\item{xvar}{Variable on the x-axis in quotes; e.g. age group}

\item{yvar}{Variable on the y-axis in quotes with the rate or number of cases}

\item{fill_color1}{Bar 1 colour; defaults to ECDC green "#65B32E"}

\item{ytitle}{y-axis title; defaults to "Rate".}
}
\description{
This function draws a barchart by age group, AER style.
Expects aggregated data.
}
\examples{
# --- Create dummy data
mydat <- data.frame(AgeGroup = c("0-25", "26-65", "65+"),
                    NumberOfCases = c(54,32,41))

# --- Plot the dummy data
plotAge(mydat,
        xvar = "AgeGroup",
        yvar = "NumberOfCases",
        ytitle = "Number of cases")
}
\keyword{AER_age_bar}
