% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAgepyramid.R
\name{plotAgepyramid}
\alias{plotAgepyramid}
\title{plotAgepyramid}
\usage{
plotAgepyramid(data, xvar, yvar, group)
}
\arguments{
\item{data}{Your data frame}

\item{xvar}{Variable with the population for each group, given in quotes.}

\item{yvar}{Variable with the agegroups for the y-axis, given in quotes.}

\item{group}{The two level grouping variable, usually gender, given in quotes.}
}
\description{
This function draws an age pyramid; improvements TBD (axis labels, selecting colours).
}
\examples{
# Create dummy data
mydat <- data.frame(Age = factor(rep(x = 1:7, times = 1),
labels = c("0-5", "6-10", "11-15",
          "16-24", "25-45", "46-64", ">65")),
Gender = rep(x = c("Female", "Male"), each = 7),
Population = rep(c(30,34,55,80,120,115,75),2))
# Plot
plotAgepyramid(data = mydat, xvar = "Population", yvar = "Age", group = "Gender")
}
\author{
Tommi Karki
}
\keyword{agepyramid}
