% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_age_gender_bar.R
\name{plotBarGrouped}
\alias{plotBarGrouped}
\title{Grouped bar graph}
\usage{
plotBarGrouped(
  .data,
  xvar = "XLabel",
  xlabel = "",
  yvar = "ZValue",
  ylabel = "",
  group = "YLabel",
  fill_color = EcdcColors(col_scale = "qual", n = length(unique(.data[[group]])))
)
}
\arguments{
\item{.data}{dataframe containing the variables to plot}

\item{xvar}{character string, name of the variable to plot on the x-axis in quotes
(default \code{"XLabel"})}

\item{xlabel}{character string, label of the x axis}

\item{yvar}{character string, name of the variable to plot on the y-axis in quotes
(default \code{"ZValue"})}

\item{ylabel}{character string, label of the y axis}

\item{group}{character string, name of the grouping variable in quotes, e.g. gender.
(default \code{"YLabel"}) .}

\item{fill_color}{vector of character strings, hexadecimal colour to use in the graph for bars;
the vector should contain the number categories in \code{"group"} variable.
(default to ECDC blue \code{"#7CBDC4"} and ECDC green \code{"#65B32E"}, see EcdcColors(col_scale = "qual", n = 2))}
}
\description{
This function draws a vertical grouped bar graph of the values of variable 'Yvar'
with the categorical variable 'Xvar' on the x-axis and grouped by 'Group' categorical variable. \cr
Expects aggregated data.
}
\examples{
# --- Create dummy data
mydat <- data.frame(Gender=c("F", "F", "M", "M"),
                    AgeGroup = c("0-65", "65+", "0-65", "65+"),
                    NumberOfCases = c(54,43,32,41))

# --- Plot the dummy data
plotBarGrouped(mydat,
              xvar = "AgeGroup",
              xlabel = "Age",
              yvar = "NumberOfCases",
              ylabel = "Number of cases",
              group = "Gender")

}
\seealso{
Global function: \code{\link{getAgeGender}}  \cr
Internal function: \code{\link{EcdcColors}} \cr
Required Packages: \code{\link{ggplot2}}
}
\keyword{bargraph}
