% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_seasonality_graph.R
\name{plotSeasonality}
\alias{plotSeasonality}
\title{AER}
\usage{
plotSeasonality(data, xvar = "TimeCode", yvar = "N",
  min4years = "Min4Years", max4years = "Max4Years",
  mean4years = "Mean4Years", year = 2016)
}
\arguments{
\item{data}{Your data frame}

\item{xvar}{Variable on the x-axis in quotes; e.g. age group}

\item{yvar}{Variable on the y-axis in quotes with the rate or number of cases}

\item{min4years}{var...}

\item{max4years}{var...}

\item{mean4years}{var..}

\item{year}{Var...}
}
\description{
This function ...
}
\keyword{seasonality}
