% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_maps.R
\name{previewMap}
\alias{previewMap}
\title{Previewing the PNG map}
\usage{
previewMap(disease, year, reportParameters, pathPNG, namePNGsuffix)
}
\arguments{
\item{disease}{character string, disease code (default \code{"SALM"}).}

\item{year}{numeric, year to produce the graph for (default \code{2016}).}

\item{reportParameters}{dataframe, dataset including the required parameters
for the graph and report production (default \code{AERparams}) (see specification
of the dataset in the package vignette with \code{browseVignettes(package = "EpiReport")})}

\item{pathPNG}{character string, full path to the folder containing the maps in PNG
(default 'maps' folder included in the package \code{system.file("maps", package = "EpiReport")})}

\item{namePNGsuffix}{character string, suffix of the PNG file name of the map
(i.e. \code{"COUNT"}, \code{"RATE"} or \code{"AGESTANDARDISED"}.)}
}
\value{
Preview
}
\description{
Function previewing the disease-specific PNG map
}
\seealso{
Global function: \code{\link{getMap}}
}
