% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SALM2016}
\alias{SALM2016}
\title{Dataset including Salmonellosis data for 2016}
\format{A data frame with 60,775 rows and 18 variables:
\describe{
  \item{HealthTopicCode}{Disease code e.g. ANTH, SALM, etc.}
  \item{MeasureLabel}{}
  \item{MeasurePopulation}{}
  \item{MeasureCode}{}
  \item{MeasureId}{}
  \item{MeasureType}{}
  \item{TimeUnit}{}
  \item{GeoLevel}{}
  \item{TimeCode}{}
  \item{GeoCode}{}
  \item{XValue}{}
  \item{XLabel}{}
  \item{YValue}{}
  \item{YLabel}{}
  \item{ZValue}{}
  \item{N}{}
  \item{NMissing}{}
  \item{NLowerResolution}{}
}}
\usage{
data(SALM2016)
}
\description{
A dataset containing the parameters required to build the AER
for Salmonellosis TESSy data
}
\keyword{datasets}
