% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_tables.R
\name{shapeECDCFlexTable}
\alias{shapeECDCFlexTable}
\title{Shaping the final table}
\usage{
shapeECDCFlexTable(ft, headers, fsize, fname, maincolor)
}
\arguments{
\item{ft}{fletable to shape into ECDC table layout}

\item{headers}{dataframe including the multiple headers to add to the flextable
(n.b. the first column should contain the names of the flextable object)}

\item{fsize}{numeric, font to use (defaut = 7)}

\item{fname}{character, font name (defaut = "Tahoma")}

\item{maincolor}{character string, hexadecial code for the header background color}
}
\value{
flextable object
}
\description{
Shaping the final table (layout)
}
