% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{tseasonalityByMS}
\alias{tseasonalityByMS}
\title{Time series graph}
\usage{
tseasonalityByMS(x)
}
\arguments{
\item{x}{dataset including required data for AER}
}
\value{
A table
}
\description{
Function drawing the first table of the AER: Number of cases by Member States over the last 5 years
}
