% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_ts_12_month_avg.R
\name{AER_ts_12_month_avg}
\alias{AER_ts_12_month_avg}
\title{AER time series with 12-month moving average}
\usage{
AER_ts_12_month_avg(data, xvar, yvar, mov_average)
}
\arguments{
\item{data}{Your data frame.}

\item{xvar}{Time variable, given in quotes.}

\item{yvar}{A variable with the number of cases for each time unit, given in quotes.}

\item{mov_average}{A variable with the moving average per each time unit.}
}
\description{
This function draws a line graph with 12-month moving average, AER style.
Expects aggregated data and pre-calculated 12-month moving average.
}
\keyword{average}
\keyword{moving}
