% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AERparams}
\alias{AERparams}
\title{Dataset including parameters for AER production}
\format{A data frame with 53 rows (corresponding to the 53 diseases) and 28 variables:
\describe{
  \item{HealthTopic}{Disease code e.g. ANTH, SALM, etc.}
  \item{DG}{}
  \item{DP}{}
  \item{Label}{}
  \item{FrequencyCategory}{}
  \item{MeasurePopulation}{}
  \item{DatePublicAtlas}{}
  \item{TableUse}{}
  \item{TableRatesLabel}{}
  \item{TableRatesNoDecimals}{}
  \item{TableASRNoDecimals}{}
  \item{AgeGenderBarGraphUse}{}
  \item{AgeGenderRatesUse}{}
  \item{AgeBarGraphUse}{}
  \item{AgeGenderBarGraphLabel}{}
  \item{AgeGenderGraphNoDecimals}{}
  \item{TSTrendGraphUse}{}
  \item{TSSeasonalityGraphUse}{}
  \item{TSSpecific}{}
  \item{MapNumbersUse}{}
  \item{MapRatesUse}{}
  \item{MapRatesNoDecimals}{}
  \item{MapASRUse}{}
  \item{MapASRNoDecimals}{}
  \item{Transmission}{}
  \item{TransmissionNoDecimals}{}
}}
\usage{
data(AERparams)
}
\description{
A dataset containing the parameters required to build the AER
for all 53 diseases included in TESSy
}
\keyword{datasets}
