% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_tables.R
\name{cleanECDCTable}
\alias{cleanECDCTable}
\title{Cleaning the final table}
\usage{
cleanECDCTable(x, Country, GeoCode)
}
\arguments{
\item{x}{dataset including required data for AER}

\item{Country}{character vector containing the correspondence table of geo code and labels}

\item{GeoCode}{character vector containing the correspondence table of geo code and labels}
}
\value{
cleaned ECDC table
}
\description{
Cleaning the final table: identifying missing reports with '-',
replacing the Member State codes with Member State names,
identifying not reporting Member States with '.'
}
