% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{cleanMeasureCode}
\alias{cleanMeasureCode}
\title{Clean the MeasureCode variable}
\usage{
cleanMeasureCode(var)
}
\arguments{
\item{var}{variable to clean}
}
\value{
cleaned variable
}
\description{
Clean the MeasureCode variable
}
