% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{filterDisease}
\alias{filterDisease}
\title{Filter disease parameters}
\usage{
filterDisease(dis, reportParameters)
}
\arguments{
\item{dis}{ccharacter string, disease code}

\item{reportParameters}{dataset of parameters for the report
(default reportParameters <- EpiReport::AERparams)}
}
\value{
dataframe with one row (from the AERparams dataframe)
corresponding to the parameters of the selected disease
}
\description{
Filter disease parameters
}
\examples{
disease <- "SALM"
reportParameters <- EpiReport::AERparams
reportParameters <- filterDisease(disease, reportParameters)
}
\seealso{
\code{\link{AERparams}}
}
