% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{filterDisease}
\alias{filterDisease}
\title{Filter disease parameters}
\usage{
filterDisease(dis, reportParameters)
}
\arguments{
\item{dis}{character string, disease code}

\item{reportParameters}{dataset of parameters for the report
(default \code{AERparams})}
}
\value{
dataframe with one row (from the \code{AERparams} dataframe)
corresponding to the parameters of the selected disease
}
\description{
Filter the table of parameters for the report on the given disease
}
\examples{
disease <- "SALM"
reportParameters <- EpiReport::AERparams
reportParameters <- filterDisease(disease, reportParameters)

}
\seealso{
\code{\link{AERparams}}
}
