% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_age_gender_bar.R
\name{getAgeGender}
\alias{getAgeGender}
\title{Get the age and gender bar graph}
\usage{
getAgeGender(x, disease = "SALM", year = 2016, reportParameters,
  geoCode = "EU_EEA31", index = 1, doc)
}
\arguments{
\item{x}{dataset}

\item{disease}{character string, disease name}

\item{year}{numeric, year to produce the report for}

\item{reportParameters}{dataset of parameters for the report}

\item{geoCode}{character string, geoCode to run the analysis on (default: EU_EEA31)}

\item{index}{figure number}

\item{doc}{Word document}
}
\value{
Word doc a ggplot2 preview
}
\description{
Function returning the plot
}
