% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{getMap}
\alias{getMap}
\title{Get the disease map}
\usage{
getMap(x, disease = "SALM", year = 2016, reportParameters, index = 1,
  pathPNG, doc)
}
\arguments{
\item{x}{dataset including required data for AER}

\item{disease}{character string, disease name}

\item{year}{numeric, year to produce the report for}

\item{reportParameters}{dataset of parameters for the report}

\item{index}{figure number}

\item{pathPNG}{character string, path to the folder containing the maps in PNG}

\item{doc}{Word document}
}
\value{
?
}
\description{
Function returning the map
}
