% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_seasonality_graph.R
\name{getSeason}
\alias{getSeason}
\title{Get the seasonal distribution graph}
\usage{
getSeason(x, disease = "SALM", year = 2016, reportParameters,
  index = 1, doc)
}
\arguments{
\item{x}{dataset}

\item{disease}{character string, disease name}

\item{year}{numeric, year to produce the report for}

\item{reportParameters}{dataset of parameters for the report}

\item{index}{figure number}

\item{doc}{Word document}
}
\value{
Word doc a ggplot2 preview
}
\description{
Function returning the plot
}
