% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{getTableByMS}
\alias{getTableByMS}
\title{Get the table}
\usage{
getTableByMS(x, disease = "Salmonellosis", year = 2016,
  reportParameters, index = 1, doc)
}
\arguments{
\item{x}{dataset including required data for AER}

\item{disease}{character string, disease name}

\item{year}{numeric, year to produce the report for}

\item{reportParameters}{dataset of parameters for the report}

\item{index}{figure number}

\item{doc}{Word document (see \code{officer} package)}
}
\value{
?
}
\description{
Function returning the table
}
