% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_tables.R
\name{getTableByMS}
\alias{getTableByMS}
\title{Get the table}
\usage{
getTableByMS(x, disease = "Salmonellosis", year = 2016,
  reportParameters, MSCode, index = 1, doc)
}
\arguments{
\item{x}{dataset including required data for AER}

\item{disease}{character string, disease name}

\item{year}{numeric, year to produce the report for}

\item{reportParameters}{dataset of parameters for the report}

\item{MSCode}{dataset containing the correspondence table of geo code and labels}

\item{index}{figure number}

\item{doc}{Word document (see \code{officer} package)}
}
\value{
Word doc or FlexTable preview
}
\description{
Function returning the table
}
