% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_report.R
\name{getTemplate}
\alias{getTemplate}
\title{Get AER Word template}
\usage{
getTemplate(output_path)
}
\arguments{
\item{output_path}{The path where to create the Word output.
Defaut location will be the current working directory (getwd())}
}
\value{
A word document
}
\description{
Function to get the standard Word template for ECDC Annual Epidemiological Report
}
\examples{
getTemplate(output_path = "C:/R/AER")
}
