% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSCode}
\alias{MSCode}
\title{Dataset correspondence table for country names and code}
\format{A data frame with 32 rows and 2 variables:
\describe{
  \item{Country}{Country name}
  \item{GeoCode}{Country code}
  \item{EUEEA}{For each country, EU or EEA depending on the country}
}}
\usage{
MSCode
}
\description{
Dataset correspondence table for country names, code and EU or EEA.
EUEEA code and name is included in a additional row in the dataframe for
the EUEEA summary of table 1 in the AER report
}
\keyword{datasets}
