% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_age_gender_bar.R
\name{plotAge}
\alias{plotAge}
\title{AER Age bar graph}
\usage{
plotAge(data, xvar = "XLabel", yvar = "YValue",
  fill_color1 = "#65B32E", ytitle = "Rate")
}
\arguments{
\item{data}{dataframe containing the variables to plot}

\item{xvar}{character string, variable on the x-axis in quotes (default "XLabel")}

\item{yvar}{character string, variable on the y-axis in quotes (default "YValue")}

\item{fill_color1}{character string, hexadecimal colour to use in the graph;
(default to ECDC green "#65B32E")}

\item{ytitle}{character string, y-axis title; (default "Rate").}
}
\description{
This function draws a barchart by age group, using the AER style.
Expects aggregated data.
}
\examples{

# --- Create dummy data
mydat <- data.frame(AgeGroup = c("0-25", "26-65", "65+"),
                    NumberOfCases = c(54,32,41))

# --- Plot the dummy data
plotAge(mydat,
        xvar = "AgeGroup",
        yvar = "NumberOfCases",
        ytitle = "Number of cases")

}
\seealso{
\code{\link{getAgeGender}} and \code{\link{plotAgeGender}}
}
\keyword{age}
\keyword{bargraph}
