% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AER_maps.R
\name{previewMap}
\alias{previewMap}
\title{Function including the PNG map}
\usage{
previewMap(disease, year, reportParameters, pathPNG, namePNGsuffix)
}
\arguments{
\item{disease}{character string, disease name}

\item{year}{numeric, year to produce the report for}

\item{reportParameters}{dataset of parameters for the report}

\item{pathPNG}{character string, path to the folder containing the maps in PNG}

\item{namePNGsuffix}{Suffix of the PNG file name of the map}
}
\value{
Preview
}
\description{
Function including the PNG map
}
