% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{toCapTitle}
\alias{toCapTitle}
\title{Capitalise the first letter}
\usage{
toCapTitle(str)
}
\arguments{
\item{str}{character string to capitalise as a title}
}
\value{
character string
}
\description{
Capitalise the first letter of a character string
}
\examples{
my_title <- "number of salmonellosis cases by age group"
toCapTitle(my_title)

}
